tela de editar

+++ src/app/pages/volunteer/my-volunteering/my-volunteering.component.ts
@@ -35,12 +35,14 @@ export class MyVolunteeringComponent implements OnInit {
       if (this.firebaseAuth.currentUser?.uid) {
         try {
           this.eventDataService.removeByUser(this.firebaseAuth.currentUser?.uid, eventToCancel)
-          this.eventDataService.removeUserOnEvent(eventToCancel.uid, this.firebaseAuth.currentUser?.email)
 
-          const element = document.getElementById(eventToCancel.uid)
-          if (element) {
-            element.remove()
-          }
+          if(this.firebaseAuth.currentUser?.email)
+            this.eventDataService.removeUserOnEvent(eventToCancel.uid, this.firebaseAuth.currentUser?.email)
+
+            const element = document.getElementById(eventToCancel.uid)
+            if (element) {
+              element.remove()
+            }
         } catch (err) {
           //
         }
